tests/manager.test.ts
import {
  assert,
  describe,
  test,
  clearStore,
  beforeAll,
  afterAll
} from "matchstick-as/assembly/index"
import { Address, BigInt } from "@graphprotocol/graph-ts"
import { AccountAdded } from "../generated/schema"
import { AccountAdded as AccountAddedEvent } from "../generated/Manager/Manager"
import { handleAccountAdded } from "../src/manager"
import { createAccountAddedEvent } from "./manager-utils"

// Tests structure (matchstick-as >=0.5.0)
// https://thegraph.com/docs/en/developer/matchstick/#tests-structure-0-5-0

describe("Describe entity assertions", () => {
  beforeAll(() => {
    let account = Address.fromString(
      "0x0000000000000000000000000000000000000001"
    )
    let ownerAccount = Address.fromString(
      "0x0000000000000000000000000000000000000001"
    )
    let newAccountAddedEvent = createAccountAddedEvent(account, ownerAccount)
    handleAccountAdded(newAccountAddedEvent)
  })

  afterAll(() => {
    clearStore()
  })

  // For more test scenarios, see:
  // https://thegraph.com/docs/en/developer/matchstick/#write-a-unit-test

  test("AccountAdded created and stored", () => {
    assert.entityCount("AccountAdded", 1)

    // 0xa16081f360e3847006db660bae1c6d1b2e17ec2a is the default address used in newMockEvent() function
    assert.fieldEquals(
      "AccountAdded",
      "0xa16081f360e3847006db660bae1c6d1b2e17ec2a-1",
      "account",
      "0x0000000000000000000000000000000000000001"
    )
    assert.fieldEquals(
      "AccountAdded",
      "0xa16081f360e3847006db660bae1c6d1b2e17ec2a-1",
      "ownerAccount",
      "0x0000000000000000000000000000000000000001"
    )

    // More assert options:
    // https://thegraph.com/docs/en/developer/matchstick/#asserts
  })
})
